requirements.txt
discord.py
python-dotenv
colorama
requests
aioconsole